Ticket #4953 - changed request_id to bigint
git-svn-id: file:///data/svn/enum/db/trunk@16065 b60c293d-5e0d-0410-87b2-fb9704ccd4b5

+++ upgrades/NEXT-notify_request.sql
@@ -4,7 +4,7 @@
 
 CREATE TABLE notify_request
 (
-    request_id INTEGER NOT NULL,
+    request_id BIGINT NOT NULL,
     message_id INTEGER REFERENCES mail_archive(id)
 );
